backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
jinja2==3.1.4
aiosmtplib==3.0.1
email-validator==2.2.0
hashlib2==1.0.0
apscheduler==3.10.4
pydantic[email]==2.9.2
